Watch Gesture Controller
Watch Gesture Controller is an Android Experiment where the Android
Wear Smartwatch sensors are used for gesture based input from the user.
Using gesture input from the smartwatch is seamless & unobtrusive for
the user, which can be handy in apps where the phone is docked away to
the TV or speaker system. In a Virtual Reality environment, using the
user’s hand gestures, it can help in creating a completely immersive
experience. We created a simple gesture based game and a gesture VR
photo library with Google Cardboard using our concept as an input. It
can be expanded into gesture based apps and games, even Android TV apps
in the future! For more: http://emberify.com/experiment